Add alias expansion after sudo calls
This aliases 'sudo' to 'sudo '. The trailing space causes the next word
to be checked for alias substitution as well, which means that we get
sudo versions of aliases for free.

In particular this is meant to make sudo cp and sudo mv expand to their
pycp/pymv versions. Hopefully this doesn't break anything else.